add react-error-boundary, add handling "we're having trouble displaying this step"
